Unblock ScreenMus::Init by calling ScreenManager::Init before gpu.
As an experiment, call ScreenManager::Init in ui/service.cc as soon
as ScreenManager::Create is called. Thus, we don't wait before
gpu is initialized (it's initialized right after ui/service and
other mojo stuff is initialized in BrowserMainLoop), and avoid
a problem, when GpuProcessHost (used when mus!=viz) has to call
to ScreenManager about initalized gpu, which starts quering of
display data, which quite tricky, because we can't set right
delegate at that time.