[DoNotCarryForward] [ozone/wayland/drm] Refactor and place GbmSurfaceless into common
This CL is a part of Wayland gpu/ui split effort, which fixes the
Ozone/Wayland and makes it functional when the gpu service is running
out of the browser process.

The patch does not bring any functionality changes, but rather
prepares Ozone/Wayland for a dmabuf based approach, which uses ozone/drm code,
to be landed in the upstream.

In particular, gbm_surfaceless.cc/h is moved into a common folder in the
ui/ozone/common/linux and renamed to gbm_surfaceless_linux.cc/h. What is more,
it is refactored in such a way that it can be used by both Wayland with
dmabuf approach and ozone/drm.

This is only compiled for wayland(downstream now) and drm backends, and
does not bring any additional load on other backends.

Bug: 820047
Change-Id: I90f7b3fd06d58e02993b4aa9461244f4d6576471