[DoNotCarryForward] Add implementation for GetRestoredBounds for DesktopWindowTreeHostPlatform
This CL introduces SetRestoredBoundsInPixels/GetRestoredBoundsInPixels
in PlatformWindow to communicate with DesktopWindowTreeHostPlatform
and makes DesktopWindowTreeHostPlatform use PlatformWindow::
GetRestoredBoundsInPixels() to get restored bounds.

When WaylandWindow is created, HandleSurfaceConfigure() is always
called without any state setting and |state_| is normal at first.
In order to know the right time when the state is changed, this
patch also updates the initial value for |state_| of WaylandWindow
to 'PLATFORM_WINDOW_STATE_NORMAL'. With this, it clears
|restored_bounds_| correctly.

Test: ozone_unittests and manual testing with ./out/ozone/chrome
--ozone-platform=wayland --in-process-gpu, see crbug.com/848131
for more details.

Issue #448

fixup! [DoNotCarryForward] Add implementation for GetRestoredBounds for DesktopWindowTreeHostPlatform