[DoNotCarryForward] Implement gbm bo wrapper in common.
This CL is a part of Wayland gpu/ui split effort, which fixes the
Ozone/Wayland and makes it functional when the gpu service is running
out of the browser process.

The patch does not bring any functionality changes, but rather
prepares Ozone/Wayland for a dmabuf based approach, which uses ozone/drm code,
to be landed in the upstream.

In particular, gbm_bo_wrapper.cc/h is now a wrapper around gbm_bo, which
is going to be used by ozone/drm and ozone/wayland.

This is only compiled for wayland(downstream now) and drm backends, and
does not bring any additional load on other backends.

TEST: compile ozone/drm.

Bug: 820047
Change-Id: I10cf2bcb990aa6de8acb67e941cc4e774747682a