Ensure default scale value of cursors is not 0.0f but 1.0f.
This patches fixes a pointer offset issues, which usually happens
with Ozone/Wayland (and can happen with Ozone/X11).

That is, on certain occasions, the pointer bitmap can have an offset,
which results in wrong visual location of the pointer. This makes
it hard to press ui elements in Chromium.

The reason of that behaviour is that default scale value of the
Cursor class was 0.0f (when it is not explicitly set to something
else), but the SearchTable (cursors_aura) method checks for the
bitmap hotpoints based on the lowest possible scale value, which
is 1.0f. Thus, there was an inconsitancy between the scale value
and the hotpoints. When the cursor had a scale value implicitly
set to 0.0f (in the ctor), 2x hotpoints were used.

Unit tests have also been added.

TEST: CursorTest.EnsureHotspotValuesWithoutScale and
CursorTest.EnsureHotspotValuesWithScale