[DoNotCarryForward] [ozone/common] Make gbm_wrapper to be compiled with system libgbm
This patch makes use of new minigbm APi names, which are aligned with
Linux libgbm.

Also, it defines certain definitions, which are different between
libgbm and minigbm.

What is more, it adds USING_MINIGBM for internal usage to be able
to distinguish between system linux gbm and minigbm.

Last but not least, drmPrimeHandleToFd is used in case of linux
system libgbm in order to be able to get mappable fds per each
plane, which is neede to add GpuMemoryBuffers support.

Bug: 869206, 578890, 820047
Change-Id: I0a743aef1c8bad2bb5792a58ab29abec45ea4585