[Ozone/X11] Add X11ScreenOzone for PlatformScreen
It implements inital PlatformScreen for Ozone/X11 with adding
X11ScreenOzone for PlatformScreen and X11DisplayFetcherOzone
to fetch display information through communicating with xrandr.

As it just handles 'AddDisplay' for now, it includes the unittest
to test if it has display or not.

X11DisplayFetcherOzone has same behavior as DesktopScreenX11,
which is used for Aura/X11, and it moved GetICCProfileForMonitor
to x11_util.cc in order to reuse code.

Change-Id: Ie24012069650ff69b528d586b5a010ebe1c7fe53